// frontend/src/types/suppliers/supplier_form.ts
export type SupplierForm = {
    id?: string

    companyName: string
    cnpj: string
    email: string
    phone: string
    contactName: string
    businessActivity: string
    password: string
    address: string
    addressNumber: string
    neighborhood: string
    city: string
    active:boolean
    state: string
    zipCode: string
}